Histology Grade and Risk Factors for Endometrial Cancer

This dataset, EndometrialCancer_df, is a data frame containing information on histology grades and associated risk factors for 79 cases of endometrial cancer. The dataset provides variables related to histological grades, pathological indices, and other clinical measures. The dataset consists of 79 observations with 4 variables.

Usage

data(EndometrialCancer_df)

Format

A data frame with 79 observations and 4 variables:

NV

Nuclear volume (integer).

PI

Pathological index (integer).

EH

Endometrial hyperplasia (numeric).

HG

Histology grade (integer).

Source

Data taken from the enrichwith package. The dataset was first analyzed in Heinze and Schemper (2002) and originally provided by Dr. E. Asseryanis from the Medical University of Vienna. The data was downloaded in .dat format from https://users.stat.ufl.edu/~aa/glm/data/, which provides datasets used in Agresti (2015).

Details

The dataset name has been kept as 'EndometrialCancer_df' to avoid confusion with other datasets in the R ecosystem. This naming convention helps distinguish this dataset as part of the OncoDataSets package and assists users in identifying its specific characteristics. The suffix '_df' indicates that the dataset is a data frame. The original content has not been modified in any way.